CIF-1444: Rename '<Account>' component to '<AccountContainer>', remove '<AccountIconText>' component (#303)
* Changes the references of account component in other files

* Merge logic of '<AccountIconText>' to '<AccountContainer>', now '<AccountTrigger>' will render icon label based on 'label' property passed in

* Update test and snapshots of '<AccountContainer>' component
