+++ react-components/src/components/AccountContainer/index.js
@@ -11,4 +11,4 @@
  *    governing permissions and limitations under the License.
  *
  ******************************************************************************/
-export { default } from './accountIconText';
+export { default } from './accountContainer';
