--- react-components/src/components/AccountIconText/accountIconText.css
@@ -1,3 +1,0 @@
-.text {
-    padding: 0 0.5rem;
-}
