+++ react-components/src/components/AccountTrigger/accountTrigger.css
@@ -3,3 +3,7 @@
     height: 3rem;
     min-width: 3rem;
 }
+
+.label {
+    padding: 0 0.5rem;
+}
